requirements.txt
fastapi
uvicorn
jieba
numpy
faiss-cpu
sentence-transformers
python-dotenv
requests
pydantic
openai
